+++ 46_APIWithTokenHomework/ViewController.m
@@ -218,7 +218,7 @@ static NSString* myVKAccountID = @"21743772";
          self.tableView.showsInfiniteScrolling = NO;
          [self.tableView.infiniteScrollingView stopAnimating];
 
-         NSLog(@"error = %@, code = %ld", [error localizedDescription], statusCode);
+         NSLog(@"error = %@, code = %ld", [error localizedDescription], (long)statusCode);
          
          UIAlertController* alertVC =
          [UIAlertController alertControllerWithTitle:@"Network error occured"
@@ -265,7 +265,7 @@ static NSString* myVKAccountID = @"21743772";
          }
          onFailure:^(NSError *error, NSInteger statusCode) {
              
-             NSLog(@"error = %@, code = %ld", [error localizedDescription], statusCode);
+             NSLog(@"error = %@, code = %ld", [error localizedDescription], (long)statusCode);
              
 
              [self.tableView.pullToRefreshView stopAnimating];
@@ -314,7 +314,7 @@ static NSString* myVKAccountID = @"21743772";
      }
      onFailure:^(NSError *error, NSInteger statusCode)
      {
-         NSLog(@"error = %@, code = %ld", [error localizedDescription], statusCode);
+         NSLog(@"error = %@, code = %ld", [error localizedDescription], (long)statusCode);
          
      }];
     
@@ -340,7 +340,7 @@ static NSString* myVKAccountID = @"21743772";
      }
      onFailure:^(NSError *error, NSInteger statusCode)
      {
-         NSLog(@"error = %@, code = %ld", [error localizedDescription], statusCode);
+         NSLog(@"error = %@, code = %ld", [error localizedDescription], (long)statusCode);
          
      }];
